Complete day 3 part 1
Rewrote a lot of the methods from the previous attempt to use a vector of numbers rather than a fixed sized tuple to allow for the longer strings in the real input.
Created this in a sub module to allow me to write it along side the existing code, and to play with modules.